refactor: remove nested components in CodeSnippetBlock
