+++ src/pages/AdvancedSearch/CodeSnippetBlock/Label/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./Label";
